refactor: split compiler.go into focused modules
- Rename names.go to constants.go (remove duplicate BacktrackLabel)
- Split compiler.go (1702 lines) into 8 focused files:
  - compiler.go: Core types and entry point (272 lines)
  - analysis.go: AST analysis utilities
  - backtracking.go: Backtracking logic
  - captures.go: Capture struct generation
  - charclass.go: Character class optimization
  - find.go: Find/FindAll generation
  - instructions.go: Instruction generators
  - pool.go: sync.Pool management

All tests pass. No functional changes.
